+++ src/components/App.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import Registry from '../utils/Registry';
 import Dashboard from './Dashboard';
-import { Router as _Router, Route, Link, browserHistory } from 'react-router'
+import { Router, Route, Link, browserHistory } from 'react-router'
 
-export default class Router extends Component {
+export default class App extends Component {
   constructor(props) {
     super(props);
   }
@@ -11,11 +11,11 @@ export default class Router extends Component {
   render() {
     let props = this.props;
     return (
-      <_Router history={browserHistory} >
+      <Router history={browserHistory} >
         <Route path="/" component={Dashboard} {...props} />
-      </_Router>   
+      </Router>   
     )
   }
 }
 
-Registry.set('Router', Router);
+Registry.set('App', App);
